Cleanup of main ChangeLog component
