+++ src/data-publication/constants/publication-changes.js
@@ -2,9 +2,10 @@ export const PRODUCT_NAMES = {
   mlar: 'Modified LAR',
   datasets: 'Datasets',
   reports: 'Reports',
-  documentation: "Documentation", 
-  apps: "HMDA Apps"
+  documentation: 'Documentation',
+  apps: 'HMDA Apps',
 }
+
 
 export const PRODUCTS = Object.keys(PRODUCT_NAMES)
 
@@ -17,11 +18,24 @@ export const CATEGORIES = {
 }
 
 
-export const PUB_CHANGELOG_URL = 'https://raw.githubusercontent.com/cfpb/hmda-frontend/master/src/data-publication/constants/publicationChangeLog.json'
+export const PUB_CHANGELOG_URL =
+  'https://raw.githubusercontent.com/cfpb/hmda-frontend/master/src/data-publication/constants/publicationChangeLog.json'
 
 
 export const DEFAULT_FILTERS = {
   type: [],
   product: [],
   keywords: [],
+}
+
+export const FILTER_OPTIONS = {
+  PRODUCT: PRODUCTS.map((product) => ({
+    value: product,
+    type: 'product',
+  })),
+
+  TYPE: Object.keys(CATEGORIES).map((type) => ({
+    value: type,
+    type: 'type',
+  })),
 }
